[refactor] : getcookie 함수 제거로인한 경로 변경
